fix: signed vs unsigned
When compiling with a recent compiler, and stronger checking, signed
vs unsigned comparison have to be avoided to pass the checks.

+++ MaxSAT.cc
@@ -351,7 +351,7 @@ void MaxSAT::printBound(int64_t bound)
 
   // print bound only, if its below the hard weight
   // FIXME: possible issue for PB instances when bound is negative; in MaxSAT bound is always positive
-  if( bound < maxsat_formula->getHardWeight() ) printf("o %" PRId64 "\n", bound);
+  if( bound < (int64_t)maxsat_formula->getHardWeight() ) printf("o %" PRId64 "\n", bound);
 }
 
 // Prints the best satisfying model. Assumes that 'model' is not empty.
